Extract CLI to top-level __main__.py
