+++ antarctica_today/__main__.py
@@ -1,0 +1,54 @@
+import click
+
+from antarctica_today.main import (
+    generate_all_plots_and_maps_main,
+    preprocessing_main,
+)
+from antarctica_today.nsidc_download_Tb_data import download_new_files
+
+
+@click.group()
+def cli():
+    """Antarctica Today."""
+    pass
+
+
+@click.option(
+    "--start-date",
+    default="2022-01-10",
+    help="TODO",
+)
+@cli.command()
+def download_tb(start_date: str):
+    """Download NSIDC-0080 granules.
+
+    The default start date is the day after the end of the .bin data available in
+    `/data/daily_melt_bin_files/` directory in this repo.
+    """
+    download_new_files(time_start=start_date)
+
+
+@cli.command()
+def preprocess():
+    """Perform pre-processing steps for Antarctica Today data.
+
+    This includes:
+        - ...
+        - ...
+    """
+    preprocessing_main()
+
+
+@cli.command()
+def process():
+    """Perform processing steps for Antarctica Today data.
+
+    This includes:
+        - ...
+        - ...
+    """
+    generate_all_plots_and_maps_main()
+
+
+if __name__ == "__main__":
+    cli()
